[FIX] Fix errors when generating volumetric nulls
The changes are the following:

-> Added a try-except statement to clean-up temporary files when an
error is raised in _vol_surrogates() function. We also modified the
cleaning-up process to reduce the likelihood that a permission error is
raised when trying to delete the temporary memmap files by now calling
._mmap.close() before unlinking the file.

-> We now convert the WindowsPath object fetched using fetch_atlas() to
a str in the mni152_to_mni152() function. This fixes an Error happening
when load_niimg(), a nibabel function, is called.

-> In the _make_surrogates() function, when genfunc() is called as
_vol_surrogates(), an error was raised because n_proc is not a valid
argument. The function call was modified so that n_proc is now a
keyword argument.

-> The generation of burt2020 nulls required version >= 0.10.0 of
brainsmash. The burt2020() function now tests for that using the
"packaging" module. This module has been added to the requirement.txt
file

-> We refactored the _vol_surrogates() function so that now, the data is
masked only when no parcellation is used. Otherwise, it is important to
note that the data should already be parcellated. As such, it is
the parcellation NIFTI image that should be used to compute the distance
matrix.  Also, the function now checks whether the data (if not
parcellated) is in the 2009cASym or 6ASym MNI152 space. If it is in
neither, then the brainmask is resampled to fit the data.

-> Fixed a bug in _vol_surrogates() where the column average was not
being computed when using a parcellation. This was causing many
bugs.

-> Fixed a bug happening when the user try to generate a single null
brainmap using the burt2018, burt2020 or moran method.

-> Modified the fetch_atlas() function so that two different T1w and
brain masks  (one for 2009cAsym and one for 6Asym) are downloaded from
OSF and added to the Bunch object when fetching the MNI152 atlases for
the 1mm and 2mm densities. We also added the name of the MNI152 version
as prefix to the map names in the Bunch object.For instance, "T1w" is
now called "2009cAsym_T1w".