[FIX] Fixed UnboundLocalError in load_data
We fixed the local UnboundLocalError that was sometimes raised when
calling load_data() with array-like `data`. This fixes, among other
things the issue discussed in PR#34.

This commit also modifies a few docstrings. One of them is the docstring
for compare_images(). We now specify that `src` and `trg` can also be
parcellated array-like data.