[FIX] Fix bug in resample_images
This commit fixes a bug in `resample_images` that would happen when a single str or os.PathLike object specifies the path to single-hemispheric `src` or `trg` images. This bug would happen when `resample_images`would try to load the gifti file of the un-transformed map. A call to the `tranforms._check_hemi` function fixes the issue.