[FIX] Fix nan_policy related bug in compare_images
This commit fixes a bug in `compare_images`. Basically, `nan` values
were dropped from `srcdata` and `trgdata` regardless of whether
`nan_policy` was set to `omit` or `propagate`. This is not what we want:
we want to keep the `nan` values when `nan_policy` is set to
`propagate`, so that the metric function can handle these values however
it wants.